+++ backend/canisters/group/impl/src/updates/c2c_claim_prize.rs
@@ -67,8 +67,10 @@ fn prepare(args: &Args, state: &mut RuntimeState) -> OCResult<PrepareResult> {
     let now_nanos = state.env.now_nanos();
     let winner = state.member_user(args.user_id);
 
+    let previous_user_ids = state.data.migrated_user_ids.previous_ids(args.user_id);
     let result = state.data.chat.reserve_prize(
         args.user_id,
+        &previous_user_ids,
         args.message_id,
         now,
         args.is_unique_person,