+++ backend/canisters/group/impl/src/updates/c2c_local_index.rs
@@ -211,7 +211,7 @@ fn process_event<F: FnOnce() -> TimestampMillis>(
             state.data.handle_bot_definition_updated(ev, **now);
         }
         LocalIndexEvent::UserIdMigrated(ev) => {
-            state.data.change_user_id(ev.old_user_id, ev.new_user_id, **now);
+            state.data.change_user_id(ev.old_user_id, ev.new_user_id, ev.principal, **now);
         }
     }