+++ backend/canisters/user_index/impl/src/updates/add_local_user_index_canister.rs
@@ -194,6 +194,7 @@ fn commit(canister_id: CanisterId, wasm_version: BuildVersion, state: &mut Runti
                 UserIndexEvent::UserIdMigrated(UserIdMigrated {
                     old_user_id,
                     new_user_id,
+                    principal: None,
                     groups: Vec::new(),
                 }),
             );